+++ docs/changes.rst
@@ -30,6 +30,7 @@ v2.4.1 (UNRELEASED)
 - Update: ``hjalmar`` after file format change.
 - Update: ``joyoftech`` after feed change.
 - Update: ``nerfnow`` to include text.
+- Update: ``pidjin`` after feed change.
 - Update: ``satw`` after site change.
 - Update: ``smbc`` after feed change.
 - Update: ``thegamercat`` after feed change.