+++ docs/changes.rst
@@ -29,6 +29,7 @@ v2.4.1 (UNRELEASED)
 - Update: ``hijinksensue`` after feed change.
 - Update: ``hjalmar`` after file format change.
 - Update: ``joyoftech`` after feed change.
+- Update: ``kiwiblitz`` after site change.
 - Update: ``nerfnow`` to include text.
 - Update: ``pidjin`` after feed change.
 - Update: ``satw`` after site change.