+++ docs/changes.rst
@@ -32,6 +32,7 @@ v2.4.1 (UNRELEASED)
 - Update: ``joyoftech`` after feed change.
 - Update: ``kiwiblitz`` after site change.
 - Update: ``lizclimo`` after feed change.
+- Update: ``lookingforgroup`` after feed change.
 - Update: ``nerfnow`` to include text.
 - Update: ``pidjin`` after feed change.
 - Update: ``satw`` after site change.